refactor: undergraduate/general-studies-requirement 리팩터링
